+++ spec/tasks/cookie_task_spec.rb
@@ -1,0 +1,54 @@
+require 'spec_helper'
+require 'rake'
+
+describe 'cookie.rake', experiments: true do
+  before do
+    Rake.application.rake_require "tasks/cookie"
+    Rake::Task.define_task(:environment)
+  end
+
+  describe 'eeny_meeny:cookie:experiment' do
+    context 'executed with an experiment id' do
+      it 'generates a cookie' do
+        expect {
+          Rake::Task['eeny_meeny:cookie:experiment'].execute(Rake::TaskArguments.new([:experiment_id],['my_page']))
+        }.to_not raise_error
+      end
+    end
+
+    context 'executed without arguments' do
+      it 'results in an error' do
+        expect {
+          Rake::Task['eeny_meeny:cookie:experiment'].execute
+        }.to raise_error(RuntimeError, "Missing 'experiment_id' parameter")
+      end
+    end
+  end
+
+  describe 'eeny_meeny:cookie:experiment_variation' do
+    context 'executed with an experiment id' do
+      it 'results in an error' do
+        expect {
+          Rake::Task['eeny_meeny:cookie:experiment_variation'].execute(Rake::TaskArguments.new([:experiment_id],['my_page']))
+        }.to raise_error(RuntimeError, "Missing 'variation_id' parameter")
+      end
+
+      context 'and a variation_id' do
+        it 'generates a cookie' do
+          expect {
+            Rake::Task['eeny_meeny:cookie:experiment_variation'].execute(Rake::TaskArguments.new([:experiment_id, :variation_id],['my_page', 'new']))
+          }.to_not raise_error
+        end
+      end
+    end
+
+    context 'executed without arguments' do
+      it 'results in an error' do
+        expect {
+          Rake::Task['eeny_meeny:cookie:experiment_variation'].execute
+        }.to raise_error(RuntimeError, "Missing 'experiment_id' parameter")
+      end
+    end
+  end
+
+end